+++ api/BusinessLogic/BudgetModels/BudgetResponse.cs
@@ -3,9 +3,9 @@
     public class BudgetResponse
     {
         public string Period { get; set; } = string.Empty;
-        public double Limit { get; set; }
-        public double Spent { get; set; }
-        public double Remaining { get; set; }
+        public decimal Limit { get; set; }
+        public decimal Spent { get; set; }
+        public decimal Remaining { get; set; }
         public int ProgressPercentage { get; set; }
     }
 }
