+++ api/DAL/Entities/Budget.cs
@@ -5,7 +5,7 @@ namespace DAL.Entities
     {
         public int Id { get; set; }
         public int UserId { get; set; }
-        public double Limit { get; set; }
+        public decimal Limit { get; set; }
         public int Year { get; set; }
         public int Month { get; set; }
         public DateTime CreatedAt { get; set; }
